requirements.txt
streamlit
pdfplumber
python-docx
epub
ebooklib
gtts
